refactor : 토큰, 유저 서비스 로직 분리
